refactor(timetable): T05 extract timetable data, undo-redo and use case
