+++ lib/components/timetable/timetable_day_list.dart
@@ -5,8 +5,8 @@ import "package:animations/animations.dart";
 import "package:collection/collection.dart";
 import "package:flutter/material.dart";
 
-import "../../isar_db/isar_timetable.dart";
-import "../../isar_db/isar_timetable_class_time.dart";
+import "../../features/timetable/models/timetable.dart";
+import "../../features/timetable/models/timetable_class_time.dart";
 import "../../pages/timetable_cell_edit_page.dart";
 import "../../sample_data.dart";
 import "../../utils/ui.dart";
