+++ lib/pages/home_tabs/tab_timetable.dart
@@ -3,7 +3,7 @@ import "package:flutter_riverpod/flutter_riverpod.dart";
 
 import "../../components/timetable/timetable.dart";
 import "../../core/pref_key.dart";
-import "../../providers/timetable_providers.dart";
+import "../../features/timetable/repositories/timetable_queries.dart";
 import "../../utils/ui.dart";
 
 class TabTimetable extends ConsumerWidget {
