+++ lib/pages/settings/timetable.dart
@@ -8,11 +8,12 @@ import "../../browser.dart";
 import "../../components/dropdown_time_picker_bottom_sheet.dart";
 import "../../core/pref_key.dart";
 import "../../features/auth/use_cases/sign_out_use_case.dart";
+import "../../features/timetable/models/timetable_class_time.dart";
+import "../../features/timetable/models/timetable_table.dart";
+import "../../features/timetable/repositories/timetable_providers.dart";
+import "../../features/timetable/repositories/timetable_queries.dart";
 import "../../infrastructure/firebase_providers.dart";
 import "../../infrastructure/firestore_providers.dart";
-import "../../isar_db/isar_timetable_class_time.dart";
-import "../../isar_db/isar_timetable_table.dart";
-import "../../providers/timetable_providers.dart";
 import "../../src/pigeons.g.dart";
 import "../../ui_components/settings_ui.dart";
 import "../../utils/ui.dart";
