+++ lib/providers/data_sync_service.dart
@@ -9,12 +9,12 @@ import "../core/pref_key.dart";
 import "../features/digestive/models/isar_digestive.dart";
 import "../features/submission/models/submission.dart";
 import "../features/submission/repositories/submission_mapper.dart";
+import "../features/timetable/models/timetable.dart";
+import "../features/timetable/models/timetable_class_time.dart";
+import "../features/timetable/models/timetable_table.dart";
 import "../infrastructure/core_providers.dart";
 import "../infrastructure/firebase_providers.dart";
 import "../infrastructure/firestore_providers.dart";
-import "../isar_db/isar_timetable.dart";
-import "../isar_db/isar_timetable_class_time.dart";
-import "../isar_db/isar_timetable_table.dart";
 import "../user_config.dart";
 import "../utils/batch_operation.dart";
 import "../utils/notifier_state_guard.dart";
